fix(cache): 캐시 용량 표시 정수 오버플로 수정
np.prod 가 Windows 에서 int32 를 돌려주어 3GB 를 넘는 캐시의 크기가 음수로
표시됐다 (4000 시나리오에서 -1.17 GB). 파이썬 정수로 곱하도록 고쳤다.
표시 전용 값이라 저장된 데이터에는 영향이 없다.

실제 캐시 4000개 x 60초를 생성해 검증했다 (129.7초, 3.0GB).
- 무결성 1500창 전수: NaN 0 / 음수 전력 0 / 전력 보존 위반 0
- 라벨 정합성 1500창 전수: OFF 인데 활성전력>0 이 0,
  ON 인데 대기전력>0 이 0, 미플러그인데 ON 이 0
- 4kW 지속부하 한도 600창 전수 준수 (최대 2661W)
- 균형 8.8:1 -> 1.4:1, 실제 추출 1.5:1
- 읽기 10,480 win/s (콜드) / 15,149 win/s (웜)

+++ src/synthesis/cache.py
@@ -149,7 +149,12 @@ def build_cache(
     }
 
     recipes: List[str] = []
-    total_bytes = sum(np.prod(s) * np.dtype(_ARRAYS[n][1]).itemsize for n, s in shapes.items())
+    # np.prod 는 Windows 에서 int32 를 돌려주므로 3GB 를 넘으면 음수가 된다.
+    # 파이썬 정수로 곱해야 한다.
+    total_bytes = sum(
+        int(np.prod(s, dtype=np.int64)) * np.dtype(_ARRAYS[n][1]).itemsize
+        for n, s in shapes.items()
+    )
     print(f"[cache] 시나리오 {n_scenarios:,}개 x {scenario_seconds:.0f}초 "
           f"= 신호 {n_scenarios * scenario_seconds / 3600:.1f}시간 | 예상 {total_bytes / 1e9:.2f} GB")
